Sovelluksessa olevaa koodia refaktoroitu.
