+++ user/models.py
@@ -2,38 +2,12 @@
 # =========
 
 from django.db import models
-#from django.contrib.auth.models import User
+from django.contrib.auth.models import User
+from django.conf import settings
 from django.http import JsonResponse
 
 # # MALLIT JOTKA MÄÄRITTELEVÄT SOVELLUKSEN TIETOKANTARAKENTEEN
 # ======================================================
-
-# Malli joka kuvaa Uutta käyttäjää sovelluksessa.
-class User(models.Model):
-  """Model representing a user in the application.
-
-  Args:
-      idNumber (BigAutoField): Unique identifier for the user
-      firstname (CharField): First name of the user
-      lastname (CharField): Last name of the user
-      phone (CharField): Phone number of the user (optional)
-      email (EmailField): Email address of the user (optional)
-      joined_date (DateField): Date when the user joined (optional)
-      slug (SlugField): Slug for URL identification
-
-  Returns:
-      str: String representation of the user (first and last name).
-  """
-  idNumber = models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')
-  firstname = models.CharField(max_length=255)
-  lastname = models.CharField(max_length=255)
-  phone = models.CharField(max_length=11, null=True)
-  email = models.EmailField(max_length=255, null=True)
-  joined_date = models.DateField(null=True)
-  slug = models.SlugField(default="", null=False)
-
-  def __str__(self):
-    return f"{self.firstname} {self.lastname}"
 
 # Malli joka kuvaa uutta tilaa sovelluksessa.
 class Space(models.Model):
@@ -53,11 +27,12 @@ class Space(models.Model):
   Returns:
       str: String representation of the space
   """
-  idNumber = models.BigAutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')
+  idNumber = models.BigAutoField(auto_created=True, primary_key=True, serialize=True, verbose_name='ID')
+#   user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE, null=True, blank=True)
   owner = models.ForeignKey(
-    'User',
+    settings.AUTH_USER_MODEL,
     on_delete=models.CASCADE,
-    related_name='spaces',
+    related_name='space',
     null=True,
     blank=True,
     verbose_name='Owner'
@@ -118,7 +93,7 @@ class Event(models.Model):
         str: String representation of the event
     """
     space = models.ForeignKey(Space, on_delete=models.CASCADE)
-    user = models.ForeignKey(User, on_delete=models.CASCADE, null=True, blank=True)
+    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE,null=True, blank=True)
     title = models.CharField(max_length=200)  # "varattu" tai "vapaa"
     start = models.DateTimeField()
     end = models.DateTimeField(blank=True, null=True)
